#20 Add Caption to ViewModels. Refactor
