+++ src/InvScanPro/Views/LocationPage.xaml
@@ -4,7 +4,7 @@
              xmlns:vm="clr-namespace:InvScanPro.ViewModels"
              x:Class="InvScanPro.Views.LocationPage"
              x:DataType="vm:LocationViewModel"
-             Title="{StaticResource Label_0014}">
+             Title="{Binding Caption}">
     <ScrollView>
         <VerticalStackLayout
             Spacing="30"
